# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Internal ERP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DashboardStats {
  suppliersCount: number;
  itemsCount: number;
  customersCount: number;
  ordersCount: number;
  invoicesCount: number;
  lowStockCount: number;
  pendingOrdersCount: number;
  totalRevenue: number;
}

export interface Supplier {
  id: number;
  name: string;
  contactName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  createdAt: string;
}

export interface CreateSupplier {
  name: string;
  contactName?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
}

export interface Item {
  id: number;
  name: string;
  description?: string | null;
  category?: string | null;
  defaultPrice?: number | null;
  createdAt: string;
}

export interface CreateItem {
  name: string;
  description?: string | null;
  category?: string | null;
  defaultPrice?: number | null;
}

export interface PriceComparison {
  id: number;
  itemId: number;
  supplierId: number;
  quotedPrice: number;
  validUntil?: string | null;
  notes?: string | null;
  createdAt: string;
  itemName?: string | null;
  supplierName?: string | null;
}

export interface CreatePriceComparison {
  itemId: number;
  supplierId: number;
  quotedPrice: number;
  validUntil?: string | null;
  notes?: string | null;
}

export interface InventoryEntry {
  id: number;
  itemId: number;
  quantityAvailable: number;
  safetyLevel: number;
  location?: string | null;
  updatedAt: string;
  itemName?: string | null;
  isLowStock: boolean;
}

export interface CreateInventoryEntry {
  itemId: number;
  quantityAvailable: number;
  safetyLevel: number;
  location?: string | null;
}

export type StockMovementMovementType =
  (typeof StockMovementMovementType)[keyof typeof StockMovementMovementType];

export const StockMovementMovementType = {
  in: "in",
  out: "out",
} as const;

export interface StockMovement {
  id: number;
  itemId: number;
  movementType: StockMovementMovementType;
  quantity: number;
  reference?: string | null;
  notes?: string | null;
  createdAt: string;
  itemName?: string | null;
}

export type CreateStockMovementMovementType =
  (typeof CreateStockMovementMovementType)[keyof typeof CreateStockMovementMovementType];

export const CreateStockMovementMovementType = {
  in: "in",
  out: "out",
} as const;

export interface CreateStockMovement {
  itemId: number;
  movementType: CreateStockMovementMovementType;
  quantity: number;
  reference?: string | null;
  notes?: string | null;
}

export interface UpdateStockMovement {
  reference?: string | null;
  notes?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  deliverySuccessRate?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateCustomer {
  name: string;
  phone?: string | null;
  email?: string | null;
  deliverySuccessRate?: number | null;
  notes?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerId: number;
  status: OrderStatus;
  totalAmount: number;
  notes?: string | null;
  createdAt: string;
  customerName?: string | null;
}

export interface OrderItem {
  id: number;
  orderId: number;
  itemId: number;
  quantity: number;
  unitPrice: number;
  itemName?: string | null;
}

export type OrderDetailStatus =
  (typeof OrderDetailStatus)[keyof typeof OrderDetailStatus];

export const OrderDetailStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderDetail {
  id: number;
  customerId: number;
  status: OrderDetailStatus;
  totalAmount: number;
  notes?: string | null;
  createdAt: string;
  customerName?: string | null;
  items: OrderItem[];
}

export type CreateOrderStatus =
  (typeof CreateOrderStatus)[keyof typeof CreateOrderStatus];

export const CreateOrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type CreateOrderItemsItem = {
  itemId: number;
  quantity: number;
  unitPrice: number;
};

export interface CreateOrder {
  customerId: number;
  status?: CreateOrderStatus;
  notes?: string | null;
  items: CreateOrderItemsItem[];
}

export type UpdateOrderStatus =
  (typeof UpdateOrderStatus)[keyof typeof UpdateOrderStatus];

export const UpdateOrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrder {
  status?: UpdateOrderStatus;
  notes?: string | null;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  issued: "issued",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface Invoice {
  id: number;
  orderId: number;
  invoiceNumber: string;
  issuedAt: string;
  total: number;
  status: InvoiceStatus;
  notes?: string | null;
  createdAt: string;
  customerName?: string | null;
}

export type CreateInvoiceStatus =
  (typeof CreateInvoiceStatus)[keyof typeof CreateInvoiceStatus];

export const CreateInvoiceStatus = {
  draft: "draft",
  issued: "issued",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface CreateInvoice {
  orderId: number;
  invoiceNumber: string;
  issuedAt: string;
  total: number;
  status: CreateInvoiceStatus;
  notes?: string | null;
}

export type ListSuppliersParams = {
  search?: string;
};

export type ListItemsParams = {
  search?: string;
  category?: string;
};

export type ListPriceComparisonsParams = {
  itemId?: number;
  supplierId?: number;
};

export type ListInventoryParams = {
  lowStock?: boolean;
};

export type ListStockMovementsParams = {
  itemId?: number;
  type?: ListStockMovementsType;
};

export type ListStockMovementsType =
  (typeof ListStockMovementsType)[keyof typeof ListStockMovementsType];

export const ListStockMovementsType = {
  in: "in",
  out: "out",
} as const;

export type ListCustomersParams = {
  search?: string;
};

export type ListOrdersParams = {
  customerId?: number;
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type ListInvoicesParams = {
  orderId?: number;
};