# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Internal ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory
      parameters:
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of inventory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryEntry"
    post:
      operationId: createInventoryEntry
      tags: [inventory]
      summary: Create inventory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryEntry"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryEntry"
  /inventory/{id}:
    get:
      operationId: getInventoryEntry
      tags: [inventory]
      summary: Get inventory entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInventoryEntry
      tags: [inventory]
      summary: Update inventory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryEntry"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryEntry"
    delete:
      operationId: deleteInventoryEntry
      tags: [inventory]
      summary: Delete inventory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Stock Movements
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Orders
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: orderId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoice"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoice"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    InventoryEntry:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        quantityAvailable:
          type: integer
        safetyLevel:
          type: integer
        location:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        itemName:
          type: string
          nullable: true
        isLowStock:
          type: boolean
      required:
        - id
        - itemId
        - quantityAvailable
        - safetyLevel
        - updatedAt
        - isLowStock
    CreateInventoryEntry:
      type: object
      properties:
        itemId:
          type: integer
        quantityAvailable:
          type: integer
        safetyLevel:
          type: integer
        location:
          type: string
          nullable: true
      required:
        - itemId
        - quantityAvailable
        - safetyLevel
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        deliverySuccessRate:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateCustomer:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        deliverySuccessRate:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        invoiceNumber:
          type: string
        issuedAt:
          type: string
          format: date-time
        total:
          type: number
        status:
          type: string
          enum: [draft, issued, paid, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        customerName:
          type: string
          nullable: true
      required:
        - id
        - orderId
        - invoiceNumber
        - issuedAt
        - total
        - status
        - createdAt
    CreateInvoice:
      type: object
      properties:
        orderId:
          type: integer
        invoiceNumber:
          type: string
        issuedAt:
          type: string
          format: date-time
        total:
          type: number
        status:
          type: string
          enum: [draft, issued, paid, cancelled]
        notes:
          type: string
          nullable: true
      required:
        - orderId
        - invoiceNumber
        - issuedAt
        - total
        - status